refactored catalog-count and catalog-count-enum

+++ flyway-db-configuration/flyway-6.5.5/sql/V0.0.4__catalog_count.sql
@@ -1,11 +1,11 @@
 CREATE TABLE IF NOT EXISTS `catalog_count` (
-  `id`                  BIGINT(8) UNSIGNED NOT NULL AUTO_INCREMENT,
-  `registration`    DATETIME               NOT NULL,
-  `catalogCountEnumId`  BIGINT(8) UNSIGNED NOT NULL,
+  `id`                  INT UNSIGNED NOT NULL AUTO_INCREMENT,
+  `registration`        DATETIME               NOT NULL,
+  `catalogCountEnumId`  INT UNSIGNED NOT NULL,
   `amount`              DOUBLE             NOT NULL,
   `details`             VARCHAR(250)        NOT NULL,
   `isDeleted`           BOOLEAN DEFAULT FALSE,
-  `branchId`            BIGINT(8) UNSIGNED NOT NULL,
+  `branchId`            INT UNSIGNED NOT NULL,
   PRIMARY KEY (`id`),
   CONSTRAINT `catalogCountEnumId_FK` FOREIGN KEY (`catalogCountEnumId`) REFERENCES `catalog_count`.`catalog_count_enum` (`id`),
   CONSTRAINT `catalogCount_branch_FK` FOREIGN KEY (`branchId`) REFERENCES `catalog_count`.`branch` (`id`)
