+++ flyway-db-configuration/flyway-6.5.5/sql/V0.0.6__Role.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS `role` (
-  `id`              BIGINT(8) UNSIGNED NOT NULL AUTO_INCREMENT,
+  `id`              INT UNSIGNED NOT NULL AUTO_INCREMENT,
   `description`     VARCHAR(64) NOT NULL,
   `registration`    DATETIME NOT NULL,
   `status`          BOOLEAN NOT NULL,
